Split object-conversion.h from native-lib.cpp
